fix: fixed spr metadata being used as a file while it's a field
